DumbBot
Closes #1
Closes #2

The BANDANA in this commit is very similar to the original version 1.3.1. just changed some functions to Java 8 and to avoid warnings.

Created a RandomBot according to the exercise in the BANDANA manual, and transformed the DumbBot to a Java file.

TournamentRunner includes some changes to run the created Bots.