Fix crash on init bandana
To fix the crash that sometimes happens when initializing parlance that leads to an agent to fail when sending its NME message, bandana is not closed and reopen every reset. Bandana will create a tournament with the max possible games and run it to the end.When the agent closes the environment,
bandana is also closed.
Added the DumbBot.jar.