DeepDip final version
This is the final version of DeepDip.
The reward function was changed to have only rewards in the end of the game.
The agent will get +5 when wins, and -5 when loses.
It will also get points for the number of Regions it owns at the end of the game.
This means the worst case is getting eliminated, it will get 0 for owned Regions and -5 for losing, total -5 reward.
The best case is winning the game +5 and to win it must control at least 5 Supply Centers, total reward >10.